Tidy up phenology class
